requirements.txt
python-telegram-bot==20.8
flask
requests
langdetect
python-dotenv
